Add identity

+++ BusinessLogic/Managers/BasketManager.cs
@@ -39,9 +39,13 @@ namespace BusinessLogic.Managers
         {
             var basket = _context.Baskets.Include(b => b.BasketProducts).Where(b => b.UserId == userId).FirstOrDefault();
 
-            _context.BasketProducts.RemoveRange(_context.BasketProducts.Where(bp => basket.BasketProducts.Select(bprod => bprod.Product).Contains(bp.Product)));
+            if(basket != null && basket.BasketProducts.Count != 0)
+            {
+                _context.BasketProducts.RemoveRange(_context.BasketProducts.Where(bp => basket.BasketProducts.Select(bprod => bprod.Product).Contains(bp.Product)));
+                _context.SaveChanges();
 
-            _context.SaveChanges();
+            }
+
         }
     }
 }
